workload

+++ entity/Workload.go
@@ -1,6 +1,6 @@
 package entity
 
 type Workload struct {
-	StaffId int `gorm:"column:id"`
-	Aggr    int `gorm:"column:aggr"`
+	StaffId int     `gorm:"column:id"`
+	Aggr    float64 `gorm:"column:aggr"`
 }
